Add FilterByOrder Method to DatabaseTD and Tests to TodoControllerSpec
This adds the filterByOrder method, which allows for filtering by specific attributes and ordering the todos in a list. Four helper methods have been added with it to ease the load on the switch statement. Additionally, more mockito tests have been added to the TodoControllerSpec class to fulfill the testing requirements.
Co-authored-by: Natasha A. Zebrev <zebre001@morris.umn.edu>